#3 feat : likes 도메인 세팅
