--- src/main/java/hongik/map/honggildong/domain/like/converter/LikeConverter.java
@@ -1,11 +1,0 @@
-package hongik.map.honggildong.domain.like.converter;
-
-import hongik.map.honggildong.domain.like.dto.LikeResponseDTO;
-import hongik.map.honggildong.domain.like.entity.Likes;
-
-public class LikeConverter {
-
-    public static LikeResponseDTO.General toGeneralDTO(Likes like){
-        return null;
-    }
-}
