+++ src/main/java/hongik/map/honggildong/domain/likes/converter/LikeConverter.java
@@ -1,0 +1,11 @@
+package hongik.map.honggildong.domain.likes.converter;
+
+import hongik.map.honggildong.domain.likes.dto.LikeResponseDTO;
+import hongik.map.honggildong.domain.likes.entity.Likes;
+
+public class LikeConverter {
+
+    public static LikeResponseDTO.General toGeneralDTO(Likes like){
+        return null;
+    }
+}
