+++ src/main/java/hongik/map/honggildong/domain/likes/entity/Likes.java
@@ -1,0 +1,19 @@
+package hongik.map.honggildong.domain.likes.entity;
+
+import hongik.map.honggildong.domain.member.entity.Member;
+import hongik.map.honggildong.domain.review.entity.Review;
+import hongik.map.honggildong.global.common.BaseEntity;
+import jakarta.persistence.*;
+
+@Entity
+public class Likes extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Review review;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Member member;
+}
